+++ cdc/sink/dispatcher/partition/default.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dispatcher
+package partition
 
 import (
 	"github.com/pingcap/tiflow/cdc/model"
